refactor: stable testids for source toggle, move constants out of components/
- Move sourcePanel.ts from components/ (component-only) up to src/ since
  it holds plain constants.
- Add data-testid="source-toggle" / "source-panel" so e2e can locate
  them without depending on DOM order or label text.
- Update the e2e helper to click via testid (no longer assumes the
  initial open state) and assert the button text via toHaveText.
